xxhsum: Split out benchmark code
Benchmark code is now in cli/xsum_bench.c.

It is still a little rough.
